requirements.txt
apache-airflow == 2.8.4
dlt == 0.4.7
pandas == 2.2.1
